Fix #1522, add printf format casts
Use (int) with %d conversions for portability

+++ modules/cfe_testcase/src/time_current_test.c
@@ -86,7 +86,7 @@ void TestGetTime(void)
     OS_GetLocalTime(&end);
 
     CFE_TIME_Print(timeBuf1, Time);
-    UtPrintf("The current time is (%d) %s", Time.Seconds, timeBuf1);
+    UtPrintf("The current time is (%ld) %s", (long)Time.Seconds, timeBuf1);
 
     difference = OS_TimeSubtract(end, start);
